requirements.txt
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
beautifulsoup4==4.12.2
requests==2.31.0
supabase==2.3.0
pydantic==2.5.3
